+++ app/src/main/java/dev/esnault/bunpyro/data/db/search/GrammarSearchResultDb.kt
@@ -5,6 +5,7 @@ data class GrammarSearchResultDb(
     val id: Long,
     val lesson: Int,
     val title: String,
+    val yomikata: String,
     val meaning: String,
     val incomplete: Boolean,
     val studied: Boolean,