+++ app/src/main/java/dev/esnault/bunpyro/data/mapper/dbtodomain/search/GrammarSearchResultMapper.kt
@@ -2,18 +2,31 @@ package dev.esnault.bunpyro.data.mapper.dbtodomain.search
 
 import dev.esnault.bunpyro.data.db.search.GrammarSearchResultDb
 import dev.esnault.bunpyro.data.mapper.IMapper
-import dev.esnault.bunpyro.domain.entities.grammar.GrammarPointOverview
+import dev.esnault.bunpyro.domain.entities.JLPT
+import dev.esnault.bunpyro.domain.entities.search.SearchGrammarOverview
 
 
-class GrammarSearchResultMapper : IMapper<GrammarSearchResultDb, GrammarPointOverview> {
+class GrammarSearchResultMapper : IMapper<GrammarSearchResultDb, SearchGrammarOverview> {
 
-    override fun map(o: GrammarSearchResultDb): GrammarPointOverview {
-        return GrammarPointOverview(
+    override fun map(o: GrammarSearchResultDb): SearchGrammarOverview {
+        return SearchGrammarOverview(
             id = o.id,
             title = o.title,
+            yomikata = o.yomikata,
             meaning = o.meaning,
             studied = o.studied,
-            incomplete = o.incomplete
+            incomplete = o.incomplete,
+            jlpt = jlptFromLesson(o.lesson)
         )
+    }
+
+    private fun jlptFromLesson(lesson: Int): JLPT {
+        return when (lesson) {
+            in 1..10 -> JLPT.N5
+            in 11..20 -> JLPT.N4
+            in 21..30 -> JLPT.N3
+            in 31..40 -> JLPT.N2
+            else -> JLPT.N1
+        }
     }
 }