+++ app/src/main/java/dev/esnault/bunpyro/domain/entities/search/SearchResult.kt
@@ -1,15 +1,13 @@
 package dev.esnault.bunpyro.domain.entities.search
 
-import dev.esnault.bunpyro.domain.entities.grammar.GrammarPointOverview
-
 
 data class SearchResult(
     /** Base query term */
     val baseQuery: String?,
     /** Query term used by the romaji -> kana search */
     val kanaQuery: String?,
-    val baseResults: List<GrammarPointOverview>,
-    val kanaResults: List<GrammarPointOverview>
+    val baseResults: List<SearchGrammarOverview>,
+    val kanaResults: List<SearchGrammarOverview>
 ) {
 
     companion object {