+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Manual sync option
 - Search headers to separate romaji search from the base search
+- JLPT tags in the search results
+- JLPT level in search results
+- Highlight query in search results
 - Navigation icons on most screens
 
 ### Fixed